server/api/requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lightgbm==4.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.26.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.4.0
plotly==5.18.0
proto-plus==1.26.0
protobuf==4.25.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
streamlit==1.29.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
xgboost==2.1.4
zipp==3.21.0
